+++ CHANGES.md
@@ -38,6 +38,11 @@
 -   #3129 Auth plugin SDK getAbsoluteUrl function should check the type of the baseUrl parameter
 -   More visual adjustment to the discourse integration
 -   Use [helm-docs v1.5.0](https://github.com/norwoodj/helm-docs/releases/tag/v1.5.0) in CI
+-   Related to #2925: Gateway deployment will auto roll on config map changes
+-   #3137 Allow UI Home Page Link URL to be configurable
+-   #3139 Make Open in National Map button Functionality Configurable & Compatible with both v7 & v8
+-   #3140 Ability to supply external CSS files to magda-web-client via Config
+-   #3138 Allow Edit Dataset Button to be replaced by External UI Plugins
 
 ## 0.0.59